[Chore] KakaoMap 버전 고정, Firebase 코드 분리
